+++ src/parser.js
@@ -544,6 +544,7 @@ var Parser = Object.extend({
 
         switch(tok.value) {
         case 'raw': return this.parseRaw();
+        case 'verbatim': return this.parseRaw('verbatim');
         case 'if':
         case 'ifAsync':
             return this.parseIf();
@@ -575,9 +576,11 @@ var Parser = Object.extend({
         return node;
     },
 
-    parseRaw: function() {
+    parseRaw: function(tagName) {
+        tagName = tagName || 'raw';
+        var endTagName = 'end' + tagName;
         // Look for upcoming raw blocks (ignore all other kinds of blocks)
-        var rawBlockRegex = /([\s\S]*?){%\s*(raw|endraw)\s*(?=%})%}/;
+        var rawBlockRegex = new RegExp('([\\s\\S]*?){%\\s*(' + tagName + '|' + endTagName + ')\\s*(?=%})%}');
         var rawLevel = 1;
         var str = '';
         var matches = null;
@@ -594,9 +597,9 @@ var Parser = Object.extend({
             var blockName = matches[2];
 
             // Adjust rawlevel
-            if(blockName === 'raw') {
+            if(blockName === tagName) {
                 rawLevel += 1;
-            } else if(blockName === 'endraw') {
+            } else if(blockName === endTagName) {
                 rawLevel -= 1;
             }