+++ browser/nunjucks.js
@@ -1,4 +1,4 @@
-/*! Browser bundle of nunjucks 2.5.2  */
+/*! Browser bundle of nunjucks 3.0.0  */
 (function webpackUniversalModuleDefinition(root, factory) {
 	if(typeof exports === 'object' && typeof module === 'object')
 		module.exports = factory();
@@ -79,7 +79,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	module.exports.lib = lib;
 	module.exports.nodes = __webpack_require__(10);
 
-	module.exports.installJinjaCompat = __webpack_require__(18);
+	module.exports.installJinjaCompat = __webpack_require__(21);
 
 	// A single instance of an environment, since this is so commonly used
 
@@ -463,6 +463,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	var builtin_loaders = __webpack_require__(14);
 	var runtime = __webpack_require__(12);
 	var globals = __webpack_require__(17);
+	var waterfall = __webpack_require__(18);
 	var Frame = runtime.Frame;
 	var Template;
 
@@ -779,7 +780,9 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	        var tmpl = new Template(src, this, opts.path);
 	        return tmpl.render(ctx, cb);
-	    }
+	    },
+
+	    waterfall: waterfall
 	});
 
 	var Context = Obj.extend({
@@ -2220,7 +2223,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	        this._compileAsyncLoop(node, frame, true);
 	    },
 
-	    _compileMacro: function(node, frame) {
+	    _compileMacro: function(node) {
 	        var args = [];
 	        var kwargs = null;
 	        var funcId = 'macro_' + this.tmpid();
@@ -2249,13 +2252,14 @@ return /******/ (function(modules) { // webpackBootstrap
 	        // arguments so support setting positional args with keywords
 	        // args and passing keyword args as positional args
 	        // (essentially default values). See runtime.js.
-	        frame = frame.push();
+	        var frame = new Frame();
 	        this.emitLines(
 	            'var ' + funcId + ' = runtime.makeMacro(',
 	            '[' + argNames.join(', ') + '], ',
 	            '[' + kwargNames.join(', ') + '], ',
 	            'function (' + realNames.join(', ') + ') {',
-	            'frame = frame.push(true);',
+	            'var callerFrame = frame;',
+	            'frame = new runtime.Frame();',
 	            'kwargs = kwargs || {};',
 	            'if (kwargs.hasOwnProperty("caller")) {',
 	            'frame.set("caller", kwargs.caller); }'
@@ -2289,8 +2293,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	          this.compile(node.body, frame);
 	        });
 
-	        frame = frame.pop();
-	        this.emitLine('frame = frame.pop();');
+	        this.emitLine('frame = callerFrame;');
 	        this.emitLine('return new runtime.SafeString(' + bufferId + ');');
 	        this.emitLine('});');
 	        this.popBufferId();
@@ -2454,14 +2457,28 @@ return /******/ (function(modules) { // webpackBootstrap
 	        var id = this.tmpid();
 	        var id2 = this.tmpid();
 
+	        this.emitLine('var tasks = [];');
+	        this.emitLine('tasks.push(');
+	        this.emitLine('function(callback) {');
 	        this.emit('env.getTemplate(');
 	        this._compileExpression(node.template, frame);
 	        this.emitLine(', false, '+this._templateName()+', ' + node.ignoreMissing + ', ' + this.makeCallback(id));
-	        this.addScopeLevel();
+	        this.emitLine('callback(null,' + id + ');});');
+	        this.emitLine('});');
 
-	        this.emitLine(id + '.render(' +
-	                      'context.getVariables(), frame, ' + this.makeCallback(id2));
-	        this.emitLine(this.buffer + ' += ' + id2);
+	        this.emitLine('tasks.push(');
+	        this.emitLine('function(template, callback){');
+	        this.emitLine('template.render(' +
+	            'context.getVariables(), frame, ' + this.makeCallback(id2));
+	        this.emitLine('callback(null,' + id2 + ');});');
+	        this.emitLine('});');
+
+	        this.emitLine('tasks.push(');
+	        this.emitLine('function(result, callback){');
+	        this.emitLine(this.buffer + ' += result;');
+	        this.emitLine('callback(null);');
+	        this.emitLine('});');
+	        this.emitLine('env.waterfall(tasks, function(){');
 	        this.addScopeLevel();
 	    },
 
@@ -2539,6 +2556,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	            this.emitFuncBegin('b_' + name);
 
 	            var tmpFrame = new Frame();
+	            this.emitLine('var frame = frame.push(true);');
 	            this.compile(block.body, tmpFrame);
 	            this.emitFuncEnd();
 	        }
@@ -2741,7 +2759,14 @@ return /******/ (function(modules) { // webpackBootstrap
 	    },
 
 	    advanceAfterVariableEnd: function() {
-	        if(!this.skip(lexer.TOKEN_VARIABLE_END)) {
+	        var tok = this.nextToken();
+
+	        if(tok && tok.type === lexer.TOKEN_VARIABLE_END) {
+	            this.dropLeadingWhitespace = tok.value.charAt(
+	                tok.value.length - this.tokens.tags.VARIABLE_END.length - 1
+	            ) === '-';
+	        } else {
+	            this.pushToken(tok);
 	            this.fail('expected variable end');
 	        }
 	    },
@@ -3156,6 +3181,7 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	        switch(tok.value) {
 	        case 'raw': return this.parseRaw();
+	        case 'verbatim': return this.parseRaw('verbatim');
 	        case 'if':
 	        case 'ifAsync':
 	            return this.parseIf();
@@ -3187,9 +3213,11 @@ return /******/ (function(modules) { // webpackBootstrap
 	        return node;
 	    },
 
-	    parseRaw: function() {
+	    parseRaw: function(tagName) {
+	        tagName = tagName || 'raw';
+	        var endTagName = 'end' + tagName;
 	        // Look for upcoming raw blocks (ignore all other kinds of blocks)
-	        var rawBlockRegex = /([\s\S]*?){%\s*(raw|endraw)\s*(?=%})%}/;
+	        var rawBlockRegex = new RegExp('([\\s\\S]*?){%\\s*(' + tagName + '|' + endTagName + ')\\s*(?=%})%}');
 	        var rawLevel = 1;
 	        var str = '';
 	        var matches = null;
@@ -3206,9 +3234,9 @@ return /******/ (function(modules) { // webpackBootstrap
 	            var blockName = matches[2];
 
 	            // Adjust rawlevel
-	            if(blockName === 'raw') {
+	            if(blockName === tagName) {
 	                rawLevel += 1;
-	            } else if(blockName === 'endraw') {
+	            } else if(blockName === endTagName) {
 	                rawLevel -= 1;
 	            }
 
@@ -3817,10 +3845,13 @@ return /******/ (function(modules) { // webpackBootstrap
 	                    this.dropLeadingWhitespace = false;
 	                }
 
-	                // Same for the succeding block start token
+	                // Same for the succeeding block start token
 	                if(nextToken &&
 	                    ((nextToken.type === lexer.TOKEN_BLOCK_START &&
 	                      nextVal.charAt(nextVal.length - 1) === '-') ||
+	                    (nextToken.type === lexer.TOKEN_VARIABLE_START &&
+	                      nextVal.charAt(this.tokens.tags.VARIABLE_START.length)
+	                        === '-') ||
 	                    (nextToken.type === lexer.TOKEN_COMMENT &&
 	                      nextVal.charAt(this.tokens.tags.COMMENT_START.length)
 	                        === '-'))) {
@@ -3844,8 +3875,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	            }
 	            else if(tok.type === lexer.TOKEN_VARIABLE_START) {
 	                var e = this.parseExpression();
-	                this.advanceAfterVariableEnd();
 	                this.dropLeadingWhitespace = false;
+	                this.advanceAfterVariableEnd();
 	                buf.push(new nodes.Output(tok.lineno, tok.colno, [e]));
 	            }
 	            else if(tok.type === lexer.TOKEN_COMMENT) {
@@ -4026,7 +4057,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	            }
 	            return token(TOKEN_BLOCK_END, tok, lineno, colno);
 	        }
-	        else if((tok = this._extractString(this.tags.VARIABLE_END))) {
+	        else if((tok = this._extractString(this.tags.VARIABLE_END)) ||
+	                (tok = this._extractString('-' + this.tags.VARIABLE_END))) {
 	            // Special check for variable end tag (see above)
 	            this.in_code = false;
 	            return token(TOKEN_VARIABLE_END, tok, lineno, colno);
@@ -4143,7 +4175,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	            this.in_code = true;
 	            return token(TOKEN_BLOCK_START, tok, lineno, colno);
 	        }
-	        else if((tok = this._extractString(this.tags.VARIABLE_START))) {
+	        else if((tok = this._extractString(this.tags.VARIABLE_START + '-')) ||
+	                (tok = this._extractString(this.tags.VARIABLE_START))) {
 	            this.in_code = true;
 	            return token(TOKEN_VARIABLE_START, tok, lineno, colno);
 	        }
@@ -5028,7 +5061,7 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	    get: function(name) {
 	        var val = this.variables[name];
-	        if(val !== undefined && val !== null) {
+	        if(val !== undefined) {
 	            return val;
 	        }
 	        return null;
@@ -5037,7 +5070,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	    lookup: function(name) {
 	        var p = this.parent;
 	        var val = this.variables[name];
-	        if(val !== undefined && val !== null) {
+	        if(val !== undefined) {
 	            return val;
 	        }
 	        return p && p.lookup(name);
@@ -5046,7 +5079,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	    resolve: function(name, forWrite) {
 	        var p = (forWrite && this.isolateWrites) ? undefined : this.parent;
 	        var val = this.variables[name];
-	        if(val !== undefined && val !== null) {
+	        if(val !== undefined) {
 	            return this;
 	        }
 	        return p && p.resolve(name);
@@ -5234,7 +5267,7 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	function contextOrFrameLookup(context, frame, name) {
 	    var val = frame.lookup(name);
-	    return (val !== undefined && val !== null) ?
+	    return (val !== undefined) ?
 	        val :
 	        context.lookup(name);
 	}
@@ -5463,8 +5496,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	        return array;
 	    },
 
-	    dump: function(obj) {
-	        return JSON.stringify(obj);
+	    dump: function(obj, spaces) {
+	        return JSON.stringify(obj, null, spaces);
 	    },
 
 	    escape: function(str) {
@@ -5581,6 +5614,13 @@ return /******/ (function(modules) { // webpackBootstrap
 	    lower: function(str) {
 	        str = normalize(str, '');
 	        return str.toLowerCase();
+	    },
+
+	    nl2br: function(str) {
+	        if (str === null || str === undefined) {
+	            return '';
+	        }
+	        return r.copySafeness(str, str.replace(/\r\n|\n/g, '<br />\n'));
 	    },
 
 	    random: function(arr) {
@@ -6194,6 +6234,275 @@ return /******/ (function(modules) { // webpackBootstrap
 
 /***/ },
 /* 18 */
+/***/ function(module, exports, __webpack_require__) {
+
+	var __WEBPACK_AMD_DEFINE_ARRAY__, __WEBPACK_AMD_DEFINE_RESULT__;/* WEBPACK VAR INJECTION */(function(setImmediate, process) {// MIT license (by Elan Shanker).
+	(function(globals) {
+	  'use strict';
+
+	  var executeSync = function(){
+	    var args = Array.prototype.slice.call(arguments);
+	    if (typeof args[0] === 'function'){
+	      args[0].apply(null, args.splice(1));
+	    }
+	  };
+
+	  var executeAsync = function(fn){
+	    if (typeof setImmediate === 'function') {
+	      setImmediate(fn);
+	    } else if (typeof process !== 'undefined' && process.nextTick) {
+	      process.nextTick(fn);
+	    } else {
+	      setTimeout(fn, 0);
+	    }
+	  };
+
+	  var makeIterator = function (tasks) {
+	    var makeCallback = function (index) {
+	      var fn = function () {
+	        if (tasks.length) {
+	          tasks[index].apply(null, arguments);
+	        }
+	        return fn.next();
+	      };
+	      fn.next = function () {
+	        return (index < tasks.length - 1) ? makeCallback(index + 1): null;
+	      };
+	      return fn;
+	    };
+	    return makeCallback(0);
+	  };
+	  
+	  var _isArray = Array.isArray || function(maybeArray){
+	    return Object.prototype.toString.call(maybeArray) === '[object Array]';
+	  };
+
+	  var waterfall = function (tasks, callback, forceAsync) {
+	    var nextTick = forceAsync ? executeAsync : executeSync;
+	    callback = callback || function () {};
+	    if (!_isArray(tasks)) {
+	      var err = new Error('First argument to waterfall must be an array of functions');
+	      return callback(err);
+	    }
+	    if (!tasks.length) {
+	      return callback();
+	    }
+	    var wrapIterator = function (iterator) {
+	      return function (err) {
+	        if (err) {
+	          callback.apply(null, arguments);
+	          callback = function () {};
+	        } else {
+	          var args = Array.prototype.slice.call(arguments, 1);
+	          var next = iterator.next();
+	          if (next) {
+	            args.push(wrapIterator(next));
+	          } else {
+	            args.push(callback);
+	          }
+	          nextTick(function () {
+	            iterator.apply(null, args);
+	          });
+	        }
+	      };
+	    };
+	    wrapIterator(makeIterator(tasks))();
+	  };
+
+	  if (true) {
+	    !(__WEBPACK_AMD_DEFINE_ARRAY__ = [], __WEBPACK_AMD_DEFINE_RESULT__ = function () {
+	      return waterfall;
+	    }.apply(exports, __WEBPACK_AMD_DEFINE_ARRAY__), __WEBPACK_AMD_DEFINE_RESULT__ !== undefined && (module.exports = __WEBPACK_AMD_DEFINE_RESULT__)); // RequireJS
+	  } else if (typeof module !== 'undefined' && module.exports) {
+	    module.exports = waterfall; // CommonJS
+	  } else {
+	    globals.waterfall = waterfall; // <script>
+	  }
+	})(this);
+
+	/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(19).setImmediate, __webpack_require__(3)))
+
+/***/ },
+/* 19 */
+/***/ function(module, exports, __webpack_require__) {
+
+	/* WEBPACK VAR INJECTION */(function(setImmediate, clearImmediate) {var nextTick = __webpack_require__(20).nextTick;
+	var apply = Function.prototype.apply;
+	var slice = Array.prototype.slice;
+	var immediateIds = {};
+	var nextImmediateId = 0;
+
+	// DOM APIs, for completeness
+
+	exports.setTimeout = function() {
+	  return new Timeout(apply.call(setTimeout, window, arguments), clearTimeout);
+	};
+	exports.setInterval = function() {
+	  return new Timeout(apply.call(setInterval, window, arguments), clearInterval);
+	};
+	exports.clearTimeout =
+	exports.clearInterval = function(timeout) { timeout.close(); };
+
+	function Timeout(id, clearFn) {
+	  this._id = id;
+	  this._clearFn = clearFn;
+	}
+	Timeout.prototype.unref = Timeout.prototype.ref = function() {};
+	Timeout.prototype.close = function() {
+	  this._clearFn.call(window, this._id);
+	};
+
+	// Does not start the time, just sets up the members needed.
+	exports.enroll = function(item, msecs) {
+	  clearTimeout(item._idleTimeoutId);
+	  item._idleTimeout = msecs;
+	};
+
+	exports.unenroll = function(item) {
+	  clearTimeout(item._idleTimeoutId);
+	  item._idleTimeout = -1;
+	};
+
+	exports._unrefActive = exports.active = function(item) {
+	  clearTimeout(item._idleTimeoutId);
+
+	  var msecs = item._idleTimeout;
+	  if (msecs >= 0) {
+	    item._idleTimeoutId = setTimeout(function onTimeout() {
+	      if (item._onTimeout)
+	        item._onTimeout();
+	    }, msecs);
+	  }
+	};
+
+	// That's not how node.js implements it but the exposed api is the same.
+	exports.setImmediate = typeof setImmediate === "function" ? setImmediate : function(fn) {
+	  var id = nextImmediateId++;
+	  var args = arguments.length < 2 ? false : slice.call(arguments, 1);
+
+	  immediateIds[id] = true;
+
+	  nextTick(function onNextTick() {
+	    if (immediateIds[id]) {
+	      // fn.call() is faster so we optimize for the common use-case
+	      // @see http://jsperf.com/call-apply-segu
+	      if (args) {
+	        fn.apply(null, args);
+	      } else {
+	        fn.call(null);
+	      }
+	      // Prevent ids from leaking
+	      exports.clearImmediate(id);
+	    }
+	  });
+
+	  return id;
+	};
+
+	exports.clearImmediate = typeof clearImmediate === "function" ? clearImmediate : function(id) {
+	  delete immediateIds[id];
+	};
+	/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(19).setImmediate, __webpack_require__(19).clearImmediate))
+
+/***/ },
+/* 20 */
+/***/ function(module, exports) {
+
+	// shim for using process in browser
+
+	var process = module.exports = {};
+	var queue = [];
+	var draining = false;
+	var currentQueue;
+	var queueIndex = -1;
+
+	function cleanUpNextTick() {
+	    draining = false;
+	    if (currentQueue.length) {
+	        queue = currentQueue.concat(queue);
+	    } else {
+	        queueIndex = -1;
+	    }
+	    if (queue.length) {
+	        drainQueue();
+	    }
+	}
+
+	function drainQueue() {
+	    if (draining) {
+	        return;
+	    }
+	    var timeout = setTimeout(cleanUpNextTick);
+	    draining = true;
+
+	    var len = queue.length;
+	    while(len) {
+	        currentQueue = queue;
+	        queue = [];
+	        while (++queueIndex < len) {
+	            if (currentQueue) {
+	                currentQueue[queueIndex].run();
+	            }
+	        }
+	        queueIndex = -1;
+	        len = queue.length;
+	    }
+	    currentQueue = null;
+	    draining = false;
+	    clearTimeout(timeout);
+	}
+
+	process.nextTick = function (fun) {
+	    var args = new Array(arguments.length - 1);
+	    if (arguments.length > 1) {
+	        for (var i = 1; i < arguments.length; i++) {
+	            args[i - 1] = arguments[i];
+	        }
+	    }
+	    queue.push(new Item(fun, args));
+	    if (queue.length === 1 && !draining) {
+	        setTimeout(drainQueue, 0);
+	    }
+	};
+
+	// v8 likes predictible objects
+	function Item(fun, array) {
+	    this.fun = fun;
+	    this.array = array;
+	}
+	Item.prototype.run = function () {
+	    this.fun.apply(null, this.array);
+	};
+	process.title = 'browser';
+	process.browser = true;
+	process.env = {};
+	process.argv = [];
+	process.version = ''; // empty string to avoid regexp issues
+	process.versions = {};
+
+	function noop() {}
+
+	process.on = noop;
+	process.addListener = noop;
+	process.once = noop;
+	process.off = noop;
+	process.removeListener = noop;
+	process.removeAllListeners = noop;
+	process.emit = noop;
+
+	process.binding = function (name) {
+	    throw new Error('process.binding is not supported');
+	};
+
+	process.cwd = function () { return '/' };
+	process.chdir = function (dir) {
+	    throw new Error('process.chdir is not supported');
+	};
+	process.umask = function() { return 0; };
+
+
+/***/ },
+/* 21 */
 /***/ function(module, exports) {
 
 	function installCompat() {
@@ -6231,6 +6540,9 @@ return /******/ (function(modules) { // webpackBootstrap
 	        throw new Error('KeyError');
 	      }
 	      return this.splice(index, 1);
+	    },
+	    append: function(element) {
+	        return this.push(element);
 	    },
 	    remove: function(element) {
 	      for (var i = 0; i < this.length; i++) {