+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Changelog
 * Fix handling of macro arg with default value which shares a name with another
   macro. Merge of [#791](https://github.com/mozilla/nunjucks/pull/791).
 
+* Add support for the spaces parameter in the dump template filter. 
+  Merge of [#868](https://github.com/mozilla/nunjucks/pull/868). 
+  Thanks Jesse Eikema
+
 
 2.5.2 (Sep 14 2016)
 ----------------